Struct rusoto_mock::MockRequestDispatcher

pub struct MockRequestDispatcher { /* fields omitted */ }

Composes mock API responses

A Default is provided which returns an successful response with an empty body

These can be constructed using either a Default implementation or with the with_status function.

Methods

impl MockRequestDispatcher

pub fn with_status(status: u16) -> MockRequestDispatcher

Returns a instance that mocks the status code that would be returned from AWS

pub fn with_dispatch_error(error: HttpDispatchError) -> MockRequestDispatcher

Mocks the service request failing with a communications error

pub fn with_body(self, body: &str) -> MockRequestDispatcher

Mocks the service response body what would be returned from AWS

pub fn with_json_body<B>(self, body: B) -> MockRequestDispatcher where
B: Serialize,

Mocks the json serialized response body what would be returned from AWS

pub fn with_request_checker<F>(self, checker: F) -> MockRequestDispatcher where
F: Fn(&SignedRequest) + Send + Sync + 'static,

Mocks the signed request checking applied to a request before sending to AWS

pub fn with_header(self, key: &str, value: &str) -> MockRequestDispatcher

Mocks a single service header that would be returned from AWS